Needed to generate a TLS cert from a CA
I could not get the docker image that UdapEd runs on to trust a self-signed cert, even if I added it as a trusted CA.  Works on Windows but not on this version of Linux.  I don't want to add code to the HttpClient to ignore the chain build failure.  I convinced myself I would not need to do it this way.  I guess that is why I did this over on the udap-dotnet project.

Another note.  I now have to re-generate all the UDAP Certificates and Intermediates to include host.docker.internal in the CRL and AIA urls so when running in Docker I can discover the static content in udap.certificates.server.devdays